Initial DataProviderExecutor
